pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kernpy"
version = "0.7.0"
authors = [
    { name = "David Rizo Valero", email = "drizo@ua.es"},
    { name = "Joan Cerveto Serrano", email = "joan.cerveto@ua.es"}
]
description = "Python Humdrum **kern and **mens utilities"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
   'antlr4-python3-runtime',
    'Pillow',
    'requests',
    'urllib3>=1.26.6',
    'tqdm'
]

[project.urls]
Homepage = "https://github.com/OMR-PRAIG-UA-ES/kernpy.git"
Documentation = "https://github.com/OMR-PRAIG-UA-ES/kernpy.git"
Issues = "https://github.com/OMR-PRAIG-UA-ES/kernpy.git"
